refactor: 채팅방 생성 함수를 유틸 함수로 분리
