+++ src/pages/connect/ConnectProfilePage.js
@@ -16,7 +16,7 @@ import {
 import { formatProfileData } from "util/formatProfileData";
 import { getMyMemberId, getRefreshToken } from "util/secureStoreUtils";
 import { useWebSocket } from "context/WebSocketContext";
-import { createSingleChatroom } from "config/api";
+import { createChatroom } from "util/createChatroom";
 
 import ConnectProfileTopBar from "@components/connect/ConnectProfileTopBar";
 import ConnectProfileBackground from "@components/connect/ConnectProfileBackground";
@@ -167,35 +167,20 @@ const ConnectProfilePage = ({ route }) => {
 		);
 	};
 
-	const isRelevantSingleChatroom = (chatroom, myMemberId, otherMemberId) => {
-		if (chatroom.chatroom_type !== "SINGLE") {
-			return false;
-		}
-		const members = chatroom.members;
-		const memberIds = members.map((member) => member.id);
-		return (
-			memberIds.includes(myMemberId) && memberIds.includes(otherMemberId)
-		);
-	};
-
 	const handleCreateSingleChatroom = async () => {
 		try {
-			const myMemberId = await getMyMemberId();
-			let chatroomInfo = chatrooms.find((chatroom) =>
-				isRelevantSingleChatroom(chatroom, myMemberId, memberId),
+			const chatroomInfo = await createChatroom(
+				memberId,
+				name,
+				chatrooms,
+				subscribeToNewChatroom,
+				token,
 			);
-
-			if (!chatroomInfo) {
-				const response = await createSingleChatroom(memberId, name);
-				chatroomInfo = response.data;
-				subscribeToNewChatroom(chatroomInfo.id, token);
-			}
 			navigation.navigate("ChatRoomPage", {
 				chatroomInfo,
 			});
 		} catch (error) {
 			Sentry.captureException(error);
-			console.log("채팅방 생성 에러:", error);
 		}
 	};
